examples/parameter_controller/requirements.txt
absl-py
six
sqlparse
dm-launchpad[tensorflow]
numpy
cloudpickle
cloud-sql-python-connector[pymysql]